Get rid of ActionEndpoint
